requirements.txt
PyYAML>=3.12
kazoo>=2.2.1.dev0
six>=1.10.0
kerberos>=1.2.5
gunicorn>=19.6.0
Flask>=0.11.1
Werkzeug>=0.11.10
MarkupSafe>=0.23
Jinja2>=2.8
itsdangerous>=0.22
click>=6.6
